requirements.txt
streamlit
tinydb
requests
pypdf
python-docx
pandas